Added Holodice.py
Bot does not answer to himself.
Dice-rolling is now possible up from a d2.
Added a feature to make roll 1d20 up to 3d20 easier.
Added coinflip command as an alternative to a d2.
Added hit dice to make it possible to 'hit a specific bodypart'.
Added a seperator to create space if neccessary.
Added a help command to get a list of all normal command.
Added a helpmaster command for a full set of all commands possible.
Added a quickfix to ensure all inputs with only a single number get processed correctly.
Added a feature to mention/ping the player the answered result is for to reduce confuison in differentiation the dice results to players.
Added a little something :P
Added an exit command to shut down the bot.
Added links to important Star Wars related websites.
Added a Theme and Themeping.
Modified the token to not be shown in source Code. It now is neccessary to input it the first time you use the bot and gets saved for future uses outside the source code.
Thx for reading, have a fabulous day.